+++ src/Armada.Runtimes/Properties/AssemblyInfo.cs
@@ -1,1 +1,2 @@
 [assembly: System.Runtime.CompilerServices.InternalsVisibleTo("Test.Shared")]
+[assembly: System.Runtime.CompilerServices.InternalsVisibleTo("Armada.Test.Runtimes")]